+++ src/grepros/ros2.py
@@ -8,7 +8,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     02.11.2021
-@modified    19.04.2022
+@modified    27.10.2022
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros.ros2
@@ -33,7 +33,7 @@ import rosidl_parser.definition
 import rosidl_runtime_py.utilities
 import yaml
 
-from . common import ConsolePrinter, Decompressor, MatchMarkers, memoize
+from . common import ConsolePrinter, MatchMarkers, memoize
 from . import rosapi
 
 
@@ -99,20 +99,19 @@ CREATE TABLE IF NOT EXISTS topics (
 );
 
 CREATE INDEX IF NOT EXISTS timestamp_idx ON messages (timestamp ASC);
+
+PRAGMA journal_mode=WAL;
+PRAGMA synchronous=NORMAL;
     """
 
 
-    def __init__(self, filename, decompress=False, progress=False, **_):
+    def __init__(self, filename, mode="a", *_, **__):
         """
-        @param   filename    bag file path to open
-        @param   decompress  decompress archived bag to file directory
-        @param   progress    show progress bar with decompression status
+        @param   filename  bag file path to open
+        @param   mode      file will be overwritten if "w"
         """
-        if Decompressor.is_compressed(filename):
-            if decompress: filename = Decompressor.decompress(filename, progress)
-            else: raise Exception("decompression not enabled")
-
         self._db     = None  # sqlite3.Connection instance
+        self._mode   = mode
         self._topics = {}    # {(topic, typename): {id, name, type}}
         self._counts = {}    # {(topic, typename, typehash): message count}
         self._qoses  = {}    # {(topic, typename): [{qos profile dict}]}
@@ -207,14 +206,16 @@ CREATE INDEX IF NOT EXISTS timestamp_idx ON messages (timestamp ASC);
         return get_message_type_hash(msg_or_type)
 
 
-    def read_messages(self, topics=None, start_time=None, end_time=None):
+    def read_messages(self, topics=None, start_time=None, end_time=None, raw=False, *_, **__):
         """
         Yields messages from the bag, optionally filtered by topic and timestamp.
 
         @param   topics      list of topics or a single topic to filter by, if at all
         @param   start_time  earliest timestamp of message to return, as UNIX timestamp
         @param   end_time    latest timestamp of message to return, as UNIX timestamp
-        @return              (topic, msg, rclpy.time.Time)
+        @param   raw         if True, then returned messages are tuples of
+                             (typename, bytes, typehash, typeclass)
+        @return              generator of (topic, message, rclpy.time.Time) tuples
         """
         self.get_topic_info()
         if not self._topics or (topics is not None and not topics):
@@ -235,21 +236,26 @@ CREATE INDEX IF NOT EXISTS timestamp_idx ON messages (timestamp ASC);
         sql += " ORDER BY timestamp"
 
         topicmap   = {v["id"]: v for v in self._topics.values()}
-        msgtypes   = {}     # {full typename: cls}
+        msgtypes   = {}  # {typename: cls}
+        topicset   = set(topics or [t for t, _ in self._topics])
         errortypes = set()  # {typename failed to instantiate, }
         for row in self._db.execute(sql, args):
             tdata = topicmap[row["topic_id"]]
-            topic, typename = tdata["name"], tdata["type"]
+            topic, typename = tdata["name"], canonical(tdata["type"])
+            if not raw and msgtypes.get(typename, typename) is None: continue # for row
+            typehash = next(h for t, n, h in self._counts if (t, n) == (topic, typename))
 
             try:
                 cls = msgtypes.get(typename) or \
                       msgtypes.setdefault(typename, get_message_class(typename))
-                msg = rclpy.serialization.deserialize_message(row["data"], cls)
+                if raw: msg = (typename, row["data"], typehash, cls)
+                else:   msg = rclpy.serialization.deserialize_message(row["data"], cls)
             except Exception as e:
                 errortypes.add(typename)
                 ConsolePrinter.warn("Error loading type %s in topic %s: %%s" %
                                     (typename, topic), e, __once=True)
-                if errortypes == set(n for _, n in self._topics):
+                if raw: msg = (typename, row["data"], typehash, None)
+                elif set(n for n, c in msgtypes.items() if c is None) == topicset:
                     break  # for row
                 continue  # for row
             errortypes.discard(typename)
@@ -295,19 +301,26 @@ CREATE INDEX IF NOT EXISTS timestamp_idx ON messages (timestamp ASC);
         """Closes the bag file."""
         if self._db:
             self._db.close()
-            self._db = None
+            self._db   = None
+            self._mode = None
 
 
     @property
     def size(self):
-        """Returns current file size."""
-        return os.path.getsize(self.filename) if os.path.isfile(self.filename) else None
+        """Returns current file size in bytes (including journaling files)."""
+        result = os.path.getsize(self.filename) if os.path.isfile(self.filename) else None
+        for suffix in ("-journal", "-wal") if result else ():
+            path = self.filename + suffix
+            result += os.path.getsize(path) if os.path.isfile(path) else 0
+        return result
 
 
     def _ensure_open(self, populate=False):
         """Opens bag database if not open, can populate schema if not present."""
         if self._db:
             return
+        if "w" == self._mode and os.path.exists(self.filename):
+            os.remove(self.filename)
         self._db = sqlite3.connect(self.filename, detect_types=sqlite3.PARSE_DECLTYPES,
                                    isolation_level=None, check_same_thread=False)
         self._db.row_factory = lambda cursor, row: dict(sqlite3.Row(cursor, row))
@@ -380,20 +393,21 @@ def canonical(typename):
     """
     is_array, bound, dimension = False, "", ""
 
-    match = re.match("sequence<(.+)>", typename)
-    if match:  # "sequence<uint8, 100>" or "sequence<uint8>"
-        is_array = True
-        typename = match.group(1)
-        match = re.match(r"([^,]+)?,\s?(\d+)", typename)
-        if match:  # sequence<uint8, 10>
+    if "<" in typename:
+        match = re.match("sequence<(.+)>", typename)
+        if match:  # "sequence<uint8, 100>" or "sequence<uint8>"
+            is_array = True
             typename = match.group(1)
-            if match.lastindex > 1: dimension = match.group(2)
+            match = re.match(r"([^,]+)?,\s?(\d+)", typename)
+            if match:  # sequence<uint8, 10>
+                typename = match.group(1)
+                if match.lastindex > 1: dimension = match.group(2)
 
-    match = re.match("(w?string)<(.+)>", typename)
-    if match:  # string<5>
-        typename, bound = match.groups()
+        match = re.match("(w?string)<(.+)>", typename)
+        if match:  # string<5>
+            typename, bound = match.groups()
 
-    if "[" in typename:  # "string<=5[<=10]" or "string<=5[10]"
+    if "[" in typename:  # "string<=5[<=10]" or "string<=5[10]" or "byte[10]" or "byte[]"
         dimension = typename[typename.index("[") + 1:typename.index("]")]
         typename, is_array = typename[:typename.index("[")], True
 
@@ -405,21 +419,6 @@ def canonical(typename):
 
     suffix = ("<=%s" % bound if bound else "") + ("[%s]" % dimension if is_array else "")
     return DDS_TYPES.get(typename, typename) + suffix
-
-
-def create_bag_reader(filename, decompress=False, progress=False, **__):
-    """
-    Returns a ROS2 bag reader with rosbag.Bag-like interface.
-
-    @param   decompress   decompress archived bag to file directory
-    @param   progress     show progress bar with decompression status
-    """
-    return Bag(filename, decompress=decompress, progress=progress)
-
-
-def create_bag_writer(filename):
-    """Returns a ROS2 bag writer with rosbag.Bag-like interface."""
-    return Bag(filename)
 
 
 def create_publisher(topic, cls_or_typename, queue_size):
@@ -462,6 +461,11 @@ def create_subscriber(topic, cls_or_typename, handler, queue_size):
     return sub
 
 
+def deserialize_message(raw, cls):
+    """Returns ROS2 message or service request/response instantiated from serialized binary."""
+    return rclpy.serialization.deserialize_message(raw, cls)
+
+
 def format_message_value(msg, name, value):
     """
     Returns a message attribute value as string.
@@ -486,6 +490,8 @@ def get_message_class(typename):
 
 def get_message_data(msg):
     """Returns ROS2 message as a serialized binary."""
+    with rosapi.TypeMeta.make(msg) as m:
+        if m.data is not None: return m.data
     return rclpy.serialization.serialize_message(msg)